trivial: Make SPVProof standalone class
